Add ConfigService to horreum client
