+++ horreum-api/src/main/java/io/hyperfoil/tools/horreum/api/services/ConfigService.java
@@ -1,6 +1,7 @@
 package io.hyperfoil.tools.horreum.api.services;
 
-import io.hyperfoil.tools.horreum.api.Version;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.quarkus.runtime.Startup;
 
 import javax.annotation.security.PermitAll;
 import javax.validation.constraints.NotNull;
@@ -9,39 +10,21 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import org.eclipse.microprofile.config.ConfigProvider;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import io.quarkus.runtime.Startup;
-
 @Startup
 @PermitAll
 @Path("/api/config")
 @Produces(MediaType.APPLICATION_JSON)
-public class ConfigService {
+public interface ConfigService {
    long startTimestamp = System.currentTimeMillis();
 
    @GET
    @Path("keycloak")
-   public KeycloakConfig keycloak() {
-      KeycloakConfig config = new KeycloakConfig();
-      config.url = getString("horreum.keycloak.url");
-      return config;
-   }
+   KeycloakConfig keycloak();
 
    @GET
    @Path("version")
-   public VersionInfo version() {
-      VersionInfo info = new VersionInfo();
-      info.version = Version.VERSION;
-      info.startTimestamp = startTimestamp;
-      return info;
-   }
+   VersionInfo version();
 
-   private String getString(String propertyName) {
-      return ConfigProvider.getConfig().getValue(propertyName, String.class);
-   }
 
    public static class VersionInfo {
       @NotNull
